+++ interactor/mock/query.go
@@ -158,3 +158,41 @@ func (c *MockQueryUsecaseQueryCall) DoAndReturn(f func(context.Context, string) 
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
+
+// QueryStream mocks base method.
+func (m *MockQueryUsecase) QueryStream(ctx context.Context, query string, fn func([]string, []bool) error) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "QueryStream", ctx, query, fn)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// QueryStream indicates an expected call of QueryStream.
+func (mr *MockQueryUsecaseMockRecorder) QueryStream(ctx, query, fn any) *MockQueryUsecaseQueryStreamCall {
+	mr.mock.ctrl.T.Helper()
+	call := mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryStream", reflect.TypeOf((*MockQueryUsecase)(nil).QueryStream), ctx, query, fn)
+	return &MockQueryUsecaseQueryStreamCall{Call: call}
+}
+
+// MockQueryUsecaseQueryStreamCall wrap *gomock.Call
+type MockQueryUsecaseQueryStreamCall struct {
+	*gomock.Call
+}
+
+// Return rewrite *gomock.Call.Return
+func (c *MockQueryUsecaseQueryStreamCall) Return(arg0 error) *MockQueryUsecaseQueryStreamCall {
+	c.Call = c.Call.Return(arg0)
+	return c
+}
+
+// Do rewrite *gomock.Call.Do
+func (c *MockQueryUsecaseQueryStreamCall) Do(f func(context.Context, string, func([]string, []bool) error) error) *MockQueryUsecaseQueryStreamCall {
+	c.Call = c.Call.Do(f)
+	return c
+}
+
+// DoAndReturn rewrite *gomock.Call.DoAndReturn
+func (c *MockQueryUsecaseQueryStreamCall) DoAndReturn(f func(context.Context, string, func([]string, []bool) error) error) *MockQueryUsecaseQueryStreamCall {
+	c.Call = c.Call.DoAndReturn(f)
+	return c
+}